Fixed issue with compilation frame
When I am in another frame, the before-advice to the function
'compile' switches to the frame holding the *compilation* buffer. But
if I hit F5 again while in the frame with the *compilation* buffer it
would open a new frame. This commit fixes that issue, and compilation
now behaves the way I want it to (mostly).

Better still would be to have the frame dismiss itself if compilation
is successful; else stay visible if compilation failed.