Added two functions: run mail server, run dbUpdates
Coded up a function to encapsulate something Dale taught me: how to
run a local mailserver using Python. This fixes the problem of the
test suite failing when I don't have a connection to the company mail
server.

The second function figures out what the name of the dbUpdates file
is and runs it in the psql buffer.