Fix: sw-tail-nfmc-logs, sw-kill-logs-meta
Fixed the problem with windows splitting twice instead of once; for
every process tailing a log file I had two windows in the frame
because (shell) opens a new window, which I suspect is new behavior
under Emacs 25.

In sw-kill-logs-meta I now call (delete-process) to spare me the
tedium of repeatedly answering "yes" to killing every process.

Threw in a hack to arrange some of the windows,
sw-arrange-nfmc-log-windows.